// NestJS/src/module/user/dto/enum.ts
export enum UserRole {
    ADMIN = 'Admin',
    CUSTOMER = 'Customer',
    OWNER = 'Owner',
}

export enum AuthType {
    LOCAL = 'Local',
    FACEBOOK = 'Facebook',
    GOOGLE = 'Google',
    APPLE = 'Apple'
}


export enum AppType {
    APP = 'App',
    BROWSER = 'Browser'
}

export enum OsType {
    ANDROID = 'Android',
    IOS = 'Ios',
    WINDOWS = 'Windows'
}